+++ src/types.ts
@@ -191,6 +191,12 @@ interface TokenRequest {
   amount: bigint
 }
 
+type SourceAssetInput =
+  | (Address | TokenSymbol)[]
+  | {
+      [chainId in number]?: (Address | TokenSymbol)[]
+    }
+
 type OwnerSignerSet =
   | {
       type: 'owner'
@@ -242,7 +248,10 @@ interface BaseTransaction {
   signers?: SignerSet
   sponsored?: boolean
   eip7702InitSignature?: Hex
+  sourceAssets?: SourceAssetInput
+  feeAsset?: Address | TokenSymbol
   settlementLayers?: SettlementLayer[]
+  lockFunds?: boolean
 }
 
 interface SameChainTransaction extends BaseTransaction {
@@ -268,6 +277,7 @@ export type {
   CallInput,
   Call,
   TokenRequest,
+  SourceAssetInput,
   OwnerSet,
   OwnableValidatorConfig,
   WebauthnValidatorConfig,