+++ src/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rhinestone/sdk
 
+## 1.0.0-alpha.32
+
+### Patch Changes
+
+- 883706c: Resource locking
+
+## 1.0.0-alpha.31
+
+### Patch Changes
+
+- 533c84f: - Selecting source assets (per chain or globally) by setting `sourceAssets`
+  - Don't pass the default token request when using same-chain settlement
+  - Choosing a fee token by setting `feeAsset`
+
 ## 1.0.0-alpha.30
 
 ### Patch Changes